Fixed handling of list arguments to deal with lists of paths.
Changed escape handling for lists to only escape the ' and " characters. Other characters are no longer discarded, so we don't have to escape path separation characters under Windows making handling the command line more useful.

The Windows list separator is now the same as the Unix separator, the comma (,) character. This is so that path names are not split into two entries due to the colon (:) character.

Test cases were removed and added appropriately.