Make the public member of the OptionsStyle read only.
This is a minor change in the API. We set the OptionsStyle to be readonly, as it's given in the static method to create an options class which parses the data. After that the OptionsStyle is never used. Therefore, there's no reason to make it settable outside of the context of this class.