Simplify the GenericList implementation
The GenericList is intended to be a wrapper only. The original implementation was inefficient by trying to also wrap not only a non-generic list, but a generic list only. As the GenericList<T> already derives from IList<T>, there's no need to wrap that, code should just use the correct type in the first place (IList<T>).

Documentation was cleaned up and expanded.

With the refactoring, moved the class to the Collections.Generic namespace where it makes sense, and add unit test cases.

All unit tests were executed and passed.

Issue: HELIOS-480