CommandLine: Internal quality improvements for MemberInfo
Code returned a "null" which was used immediately after which could lead to a NullReferenceException. Added a test case to ensure this code path isn't reached, and changed to raise an InternalApplicationException instead, which would indicate an error in refactoring the code if it occurred.

The problem was the checking of a MemberInfo type, if it is a PropertyInfo or a FieldInfo. If it was neither, then "null" was returned. It must be one or the other when scanning the attributes. Confirmed through a test case that checks what happens if the attribute is assigned to an Event.

Issue: HELIOS-550