CommandLine: Allow Options properties to use a IList<T>
Options that were defined type 'List<T>' worked because System.Collections.Generic.List<T> also implemented IList. But System.Collections.Generic.IList<T> doesn't. Thus during reflection it would think that this is not a list and try and assign a variable to it directly resulting in an error.

The fix expands the logic so that not only is IList allowed (as originally), but so is IList<T> and ICollection<T>. Further, the T doesn't need to be any longer of type "string", it can be anything that can be converted from a string to a type, like integers, the same way that non-list options work.

The OptionData class was refactored out to be a new class OptionMember to keep code readable.

Issue: HELIOS-1506