CommandLine: Treat OptionArgumentsAttribute the same as any other List option (but for strings)
Logic introduced with bb9b64a ("CommandLine: Allow Options properties to use a IList<T>") allowed to treat fields/properties defining a collection to be ICollection<T> or IList. The [OptionArgument] attribute is not much different in concept, so reuse the same code and make this field more flexible for the programmer.

The one constriction that still remains is that the arguments list must be of type "string" or "object", as it is always assigned a string, and not converted while parsing.

This allows removal of the "GenericList" which was a wrapper, and not much value.

Issue: HELIOS-1521