CommandLine: Fix minor warnings
IDE0044: Apply the readonly attribute where possible. Some instances use reflection to set fields, and these should not be marked as `readonly`.
IDE0056: Use index operator is a .NET Core only feature, so is ignored
IDE0057: Use range operator is a .NET Core only feature, so is ignored
CA1825: Avoid unnecessary zero-length array allocations is ignored for Test Projects as it is not necessary
NUnit1028: Mark an internal method as private, as only test cases should be public

Issue: DOTNET-360