CommandLine: Relax non-option arguments on the command line for Unix
Arguments which aren't required for options are parsed and pushed to the end of the argument list while parsing arguments. This internal reordering now means it is possible to give arguments at the beginning of the command line, it no longer has to be after all the options are provided.

Issue: DOTNET-929