CommandLine: Strengthen checks for redirection on .NET 4.0
Getting the console width when not on a console is undefined. On Windows, we'd get a `System.IO.IOException`. It's unclear what would happen on non-Windows platforms in all cases, so wrap around an exception handler.

This code doesn't apply to Windows, as here we use Win32 API directly to test which avoids this exception occurring.

Issue: DOTNET-926