CommandLine: Annotate with SupportedOSPlatform
Add extra annotations to further restrict the supported platforms based on Platform Invoke. This further improves the robustness of the library and incorrect usage for cross-platform support.

The P/Invoke is only used in .NET 4.x, so it will not show up in any diagnostics. It was reviewed manually.

Issue: DOTNET-938